refactor: create different files
