+++ packages/blade/src/components/Menu/VisualSubComponents/MenuDivider.web.tsx
@@ -1,0 +1,7 @@
+import { Divider } from '~components/Divider';
+
+const MenuDivider = (): React.ReactElement => {
+  return <Divider marginLeft="-8px" marginY="spacing.1" marginRight="-8px" />;
+};
+
+export { MenuDivider };
