+++ packages/blade/src/components/Menu/index.ts
@@ -1,0 +1,5 @@
+export { Menu } from './Menu';
+export { MenuItem } from './MenuItem';
+export { MenuOverlay } from './MenuOverlay';
+export { MenuDivider } from './VisualSubComponents/MenuDivider';
+export type { MenuProps, MenuItemProps } from './types';
