+++ packages/blade/src/components/Menu/useMenu.ts
@@ -1,0 +1,133 @@
+import React from 'react';
+import {
+  autoUpdate,
+  flip,
+  offset,
+  safePolygon,
+  shift,
+  useClick,
+  useDismiss,
+  useFloating,
+  useFloatingNodeId,
+  useFloatingParentNodeId,
+  useFloatingTree,
+  useHover,
+  useInteractions,
+  useListNavigation,
+  useListItem,
+  useRole,
+} from '@floating-ui/react';
+import type { MenuContextType, UseFloatingMenuProps, UseFloatingMenuReturnType } from './types';
+
+const MenuContext = React.createContext<MenuContextType>({
+  getItemProps: () => ({}),
+  // eslint-disable-next-line @typescript-eslint/no-empty-function
+  setHasFocusInside: () => {},
+  isOpen: false,
+});
+
+const useMenu = (): MenuContextType => {
+  const contextValue = React.useContext(MenuContext);
+  return contextValue;
+};
+
+const useFloatingMenuSetup = ({ elementsRef }: UseFloatingMenuProps): UseFloatingMenuReturnType => {
+  const [isOpen, setIsOpen] = React.useState(false);
+  const [activeIndex, setActiveIndex] = React.useState<number | null>(null);
+
+  const tree = useFloatingTree();
+  const nodeId = useFloatingNodeId();
+  const parentId = useFloatingParentNodeId();
+  const item = useListItem();
+
+  const isNested = parentId != null;
+
+  const { floatingStyles, refs, context } = useFloating<HTMLButtonElement>({
+    nodeId,
+    open: isOpen,
+    onOpenChange: setIsOpen,
+    placement: isNested ? 'right-start' : 'bottom-start',
+    middleware: [
+      offset({ mainAxis: isNested ? 12 : 0, alignmentAxis: isNested ? -16 : 0 }),
+      flip(),
+      shift(),
+    ],
+    whileElementsMounted: autoUpdate,
+  });
+
+  const hover = useHover(context, {
+    enabled: isNested,
+    delay: { open: 75 },
+    handleClose: safePolygon({ blockPointerEvents: true }),
+  });
+  const click = useClick(context, {
+    event: 'mousedown',
+    toggle: !isNested,
+    ignoreMouse: isNested,
+  });
+  const role = useRole(context, { role: 'menu' });
+  const dismiss = useDismiss(context, { bubbles: true });
+  const listNavigation = useListNavigation(context, {
+    listRef: elementsRef,
+    activeIndex,
+    nested: isNested,
+    onNavigate: setActiveIndex,
+    focusItemOnHover: false,
+  });
+
+  const { getReferenceProps, getFloatingProps, getItemProps } = useInteractions([
+    hover,
+    click,
+    role,
+    dismiss,
+    listNavigation,
+  ]);
+
+  // Event emitter allows you to communicate across tree components.
+  // This effect closes all menus when an item gets clicked anywhere
+  // in the tree.
+  React.useEffect(() => {
+    if (!tree) return;
+
+    const handleTreeClick = (): void => {
+      setIsOpen(false);
+    };
+
+    const onSubMenuOpen = (event: { nodeId: string; parentId: string }): void => {
+      if (event.nodeId !== nodeId && event.parentId === parentId) {
+        setIsOpen(false);
+      }
+    };
+
+    tree.events.on('click', handleTreeClick);
+    tree.events.on('menuopen', onSubMenuOpen);
+
+    // eslint-disable-next-line consistent-return
+    return () => {
+      // Cleanup
+      tree.events.off('click', handleTreeClick);
+      tree.events.off('menuopen', onSubMenuOpen);
+    };
+  }, [tree, nodeId, parentId]);
+
+  React.useEffect(() => {
+    if (isOpen && tree) {
+      tree.events.emit('menuopen', { parentId, nodeId });
+    }
+  }, [tree, isOpen, nodeId, parentId]);
+
+  return {
+    getReferenceProps,
+    getFloatingProps,
+    getItemProps,
+    item,
+    context,
+    nodeId,
+    isOpen,
+    floatingStyles,
+    refs,
+    isNested,
+  };
+};
+
+export { MenuContext, useMenu, useFloatingMenuSetup };
